refactor: cleanup snap tests
