+++ core/tests/test_runner/spec_test_runner.rs
@@ -1,1 +1,6 @@
 pub struct SpecTestRunner {}
+
+#[macro_export]
+macro_rules! run_spec_test {
+    ($spec_test:ident) => {};
+}
