+++ core/tests/test_runner/test_file.rs
@@ -1,0 +1,35 @@
+use serde::{Deserialize, Serialize};
+
+// TODO: replace with Config?
+#[derive(Deserialize, Clone, Serialize)]
+pub struct Preamble {
+    pub title: Option<String>,
+    pub author: Option<String>,
+    pub lang: Option<String>,
+}
+
+#[derive(Deserialize)]
+pub struct TestOutputs {
+    pub html: Option<String>,
+    pub um: Option<String>,
+}
+
+#[derive(Deserialize)]
+pub struct Test {
+    pub name: String,
+    pub description: Option<String>,
+    pub preamble: Option<Preamble>,
+    pub input: String,
+
+    #[serde(flatten)]
+    pub outputs: TestOutputs,
+}
+
+#[derive(Deserialize)]
+pub struct TestFile {
+    pub spec: String,
+    pub name: String,
+    pub description: Option<String>,
+    pub preamble: Option<Preamble>,
+    pub tests: Vec<Test>,
+}
